--- mtginventory/inventory/static/css/button.css
@@ -1,7 +1,0 @@
-.add{
-  border-radius: 50%;
-  font-size: 36px;
-  right: -1;
-  bottom: ;
-  position: sticky;
-}
