--- mtginventory/inventory/templates/inventory/index.html
@@ -1,46 +1,0 @@
-{% extends 'inventory/base.html' %}
-
-{% block title %}Home{% endblock %}
-{% load fontawesome %}
-{% load bootstrap4 %}
-
-{% block content %}
-        <div class="container-fluid mt-3 table-responsive">
-            <table class="table table-bordered table-hover table-dark">
-                <thead>
-                    {% for i in header %}
-                        <th scope="col">{{ i }}</th>
-                    {% endfor %}    
-                </thead>
-                <tbody>
-                    {% for row in rows %}
-                    <tr>
-                        <th scope="row"><span class="mx-auto">{{row.id}}</span></th>
-                        <td><span class="mx-auto">{{row.name}}</span></td>
-                        <td><span class="mx-auto">{{row.set_name_id}}</span></td>
-                        <td><span class="mx-auto">{{row.cmc}}</span></td>
-                        <td><span class="mx-auto">{{row.num_of_cards}}</span></td>
-                        <td><span class="mx-auto">{{row.state_of_card_id}}</span></td>
-                        <td><span class="mx-auto">{{row.owner_id}}</span></td>
-                        <td><span class="mx-auto">{{row.colour_id}}</span></td>
-                    <tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-            <button class='btn btn-dark add'>{% fontawesome_icon 'plus' %}</button>
-            <nav aria-label="Page navigation example">
-                <ul class="pagination justify-content-center">
-                    {% if row.has_previous %}
-                    <li class="page-item disabled">
-                        <a class="page-link" href="?page={{ rows.previous_page_number }}" tabindex="-1" aria-disabled="true">Previous</a>
-                    </li>
-                    {% endif %}
-                      {% if rows.has_next %}
-                    <li class="page-item">
-                    <a class="page-link" href="?page={{ rows.next_page_number }}">Next</a>
-                    </li>
-                    {% endif %}
-                </ul>
-            </nav>
-        </div>
-{% endblock %}
