+++ mtginventory_project/inventory/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class InventoryConfig(AppConfig):
+    name = 'inventory'
