+++ mtginventory_project/inventory/templates/inventory/base.html
@@ -1,0 +1,33 @@
+<html lang="en">
+{% load bootstrap4 %}
+{% load fontawesome %}
+{% load static %}
+<head>
+    <title>SF Card Collection Database</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    
+    {% fontawesome_stylesheet %}
+    {% bootstrap_css %}
+    {% bootstrap_javascript jquery='full' %}
+    {% bootstrap_messages %}
+    <link rel="stylesheet" href="{% static 'css/button.css' %}">
+    </head>
+<body>
+
+<nav class="navbar navbar-expand-md navbar-dark bg-dark">
+    <div class="container-fluid">
+        <div class="navbar-header">
+            <a class="navbar-brand" href="{% url 'index' %}">Online Card Binder</a>
+        </div>
+    </div>
+</nav>
+<main class="container-fluid">
+    {% block content %}
+    {% endblock %}
+</main>
+</body>
+<script>
+    $('ul.nav > li > a[href="' + document.location.pathname + '"]').parent().addClass('active');
+</script>
+</html>
